% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refresh.R
\name{covid19sf_refresh}
\alias{covid19sf_refresh}
\title{Refreshing the covid19sf Pakcage Datasets}
\usage{
covid19sf_refresh(force = FALSE)
}
\arguments{
\item{force}{A boolean, if set to TRUE will update the package if new data is available automatically}
}
\description{
The function enables to keep the package datasets with most recent data available on the package main repository.
The main repository is refreshed on a daily basis.
}
