% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19sf_geo}
\alias{covid19sf_geo}
\title{San Francisco COVID-19 Cases and Deaths Summarized by Geography}
\format{
An object class sf and data.frame with 8 variables.
\describe{
  \item{rate}{The rate of cases in the area, calculated as (count/acs_population) * 10000 which is a rate per 10,000 residents}
  \item{count}{The count of cases in the area}
  \item{deaths}{The number of cases in the area }
  \item{last_updated}{Last update of the data in POSIXc forecast)}
  \item{id}{area id}
  \item{area_type}{Area type}
  \item{acs_population}{The population from the latest 5-year estimates from the American Community Survey (2014-2018))}
  \item{geometry}{The area polygon data)}
  }
}
\source{
San Francisco, Department of Public Health - Population Health Division through San Francisco Opne Data protal \href{https://datasf.org/opendata/}{website}.
}
\usage{
covid19sf_geo
}
\description{
Medical provider confirmed COVID-19 cases and confirmed COVID-19 related deaths
in San Francisco, CA aggregated by several different geographic areas and normalized by
2018 American Community Survey (ACS) 5-year estimates for population data to calculate rate per 10,000 residents.
More information about the data available \href{https://data.sfgov.org/COVID-19/COVID-19-Cases-and-Deaths-Summarized-by-Geography/tpyr-dvnc}{here}
}
\details{
The dataset contains a summary of covid19 cases in San Francisco by geographic area
}
\examples{
data(covid19sf_geo)

head(covid19sf_geo)

library(sf)
# Ploting SF Covid19 map using base plot function
plot(covid19sf_geo[, c("count", "geometry")])
plot(covid19sf_geo[, c("rate", "geometry")])
}
\keyword{COVID19}
\keyword{datasets}
\keyword{geo}
\keyword{map}
\keyword{summary}
